// frontend/src/shared/api/tracking/model/showRef.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * defShows Tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { ShowRefAiringStatus } from './showRefAiringStatus';

export interface ShowRef {
  /**
     * Catalog seasons excluding specials
     * @minimum 0
     */
  season_count?: number;
  id: number;
  tmdb_id: number;
  title: string;
  original_title?: string;
  poster_url?: string;
  airing_status: ShowRefAiringStatus;
  /** @nullable */
  next_episode_air_date?: string | null;
  vote_average?: number;
  vote_count?: number;
}