// frontend/src/shared/api/tracking/model/trackedShow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * defShows Tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { Progress } from './progress';
import type { ShowRef } from './showRef';
import type { UserShow } from './userShow';

export interface TrackedShow {
  user_show: UserShow;
  show: ShowRef;
  progress: Progress;
}